pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jobplanner"
version = "0.1.0"
description = "Automated resume tailoring pipeline — paste a JD, get a tailored ATS-compliant PDF"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.0",
    "anthropic>=0.40",
    "openai>=1.0",
    "pymupdf>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
]
web = [
    "streamlit>=1.30",
]

[project.scripts]
jobplanner = "jobplanner.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/jobplanner"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
